Fix logic in insertAfterMaxWidthNode() function.
Was completely blind what happens for the first node. The first node
was the last node in my test data, although it should have been in TOP3.

The error was to check next node's null value in a loop's precondition
statement, rather than inside the loop. Putting the check inside the
loop lets us to insert the node after, if the becoming (next) node will
be null. This fixes the order of nodes.